+++ src/main.py
@@ -9,6 +9,7 @@ from config import DBSettings
 from dependencies.containers import RepositoriesContainer
 from storage.sqlalchemy.client import SqlAlchemyAsync
 from web.routers import auth_router, user_router, jobs_router
+from web.routers.response import router as response_router
 
 env_file_name = ".env." + os.environ.get("STAGE", "dev")
 env_file_path = Path(__file__).parent.resolve() / env_file_name
@@ -18,7 +19,6 @@ def create_app():
     repo_container = RepositoriesContainer()
     settings = DBSettings(_env_file=env_file_path)
 
-
     repo_container.db.override(
         providers.Factory(
             SqlAlchemyAsync,
@@ -26,13 +26,13 @@ def create_app():
         ),
     )
 
-
     app = FastAPI()
     app.container = repo_container
 
     app.include_router(auth_router)
     app.include_router(user_router)
     app.include_router(jobs_router)
+    app.include_router(response_router)
 
     return app
