+++ src/web/routers/jobs.py
@@ -1,3 +1,5 @@
+from typing import Any, Coroutine
+
 from fastapi import APIRouter, Depends, HTTPException, status
 from dependency_injector.wiring import Provide, inject
 from models import Response, User, Job  # Импортируйте свои модели
@@ -93,7 +95,7 @@ async def update_job(
     job_id: int,
     job_update_dto: JobCreateSchema,
     job_repository: JobRepository = Depends(Provide[RepositoriesContainer.job_repository])
-) -> JobSchema:
+) -> Job:
     """
     Обновляет запись о вакансии по заданному идентификатору.
 
@@ -140,42 +142,42 @@ async def delete_job(
     except Exception as e:
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"Error deleting job: {str(e)}")
 
-@router.post("/{job_id}/responses", response_model=ResponseSchema, status_code=status.HTTP_201_CREATED)
-@inject
-async def create_response(
-    job_id: int,
-    response_data: ResponseSchema,
-    current_user: User = Depends(get_current_user),
-    response_repository: ResponseRepository = Depends(Provide[RepositoriesContainer.response_repository]),
-    job_repository: JobRepository = Depends(Provide[RepositoriesContainer.job_repository])
-):
-    """
-    Создает отклик на вакансию.
-
-    Args:
-        job_id (int): Идентификатор вакансии.
-        response_data (ResponseSchema): DTO с данными отклика.
-        current_user (User): Текущий пользователь.
-        response_repository (ResponseRepository): Репозиторий для работы с откликами.
-        job_repository (JobRepository): Репозиторий для работы с вакансиями.
-
-    Returns:
-        Response: Созданный отклик на вакансию.
-
-    Raises:
-        HTTPException: Если вакансия не найдена.
-    """
-    job = await job_repository.retrieve(job_id)
-
-    if job is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Job not found")
-
-    new_response = Response(
-        user_id=current_user.id,
-        job_id=job_id,
-        message=response_data.message
-    )
-
-    await response_repository.create(new_response)
-
-    return new_response
+# @router.post("/{job_id}/responses", response_model=ResponseSchema, status_code=status.HTTP_201_CREATED)
+# @inject
+# async def create_response(
+#     job_id: int,
+#     response_data: ResponseSchema,
+#     current_user: User = Depends(get_current_user),
+#     response_repository: ResponseRepository = Depends(Provide[RepositoriesContainer.response_repository]),
+#     job_repository: JobRepository = Depends(Provide[RepositoriesContainer.job_repository])
+# ):
+#     """
+#     Создает отклик на вакансию.
+#
+#     Args:
+#         job_id (int): Идентификатор вакансии.
+#         response_data (ResponseSchema): DTO с данными отклика.
+#         current_user (User): Текущий пользователь.
+#         response_repository (ResponseRepository): Репозиторий для работы с откликами.
+#         job_repository (JobRepository): Репозиторий для работы с вакансиями.
+#
+#     Returns:
+#         Response: Созданный отклик на вакансию.
+#
+#     Raises:
+#         HTTPException: Если вакансия не найдена.
+#     """
+#     job = await job_repository.retrieve(job_id)
+#
+#     if job is None:
+#         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Job not found")
+#
+#     new_response = Response(
+#         user_id=current_user.id,
+#         job_id=job_id,
+#         message=response_data.message
+#     )
+#
+#     await response_repository.create(new_response)
+#
+#     return new_response
