+++ src/web/schemas/response.py
@@ -1,6 +1,8 @@
 from pydantic import BaseModel
 
 class ResponseSchema(BaseModel):
+    user_id: int
+    job_id: int
     message: str
 
     class Config:
